Allow float values for dash parameter
- Parse float values
- Scale dash values with a fixed scale factor "20" (In accordance with JS) before filling the dash
texture
- Make sure to adjust texture coordinate to account for the texture data
- fixes #951

+++ core/src/scene/sceneLoader.cpp
@@ -792,9 +792,9 @@ void SceneLoader::loadStyleProps(Style& style, Node styleNode, const std::shared
     if (Node dashNode = styleNode["dash"]) {
         if (auto polylineStyle = dynamic_cast<PolylineStyle*>(&style)) {
             if (dashNode.IsSequence()) {
-                std::vector<int> dashValues;
+                std::vector<float> dashValues;
                 for (auto dashValue : dashNode) {
-                    dashValues.push_back(dashValue.as<int>());
+                    dashValues.push_back(dashValue.as<float>());
                 }
                 polylineStyle->setDashArray(dashValues);
                 polylineStyle->setTexCoordsGeneration(true);
